* deprecate ToolsVersionLoaderProtocol which had no functional use (redundant abstraction) * deprecate ToolsVersionLoader and refactor it as ToolsVersionParser, a stateless utility * rename and refactor the Manfeist::path extension to ManifestLoader::findManiofest and clarify its expected usage and argument
